chore: NO_SUCH_USER 에서 NO_SUCH_MEMBER 으로 변경
